@@ -1,17 +1,17 @@
 ﻿@typeparam TItem
 
-@if (items == null)
+@if (_items is null)
 {
     @Loading
 }
-else if (!items.Any())
+else if (!_items.Any())
 {
     @Empty
 }
 else
 {
     <div class="list-group @ListGroupClass">
-        @foreach (var item in items)
+        @foreach (var item in _items)
         {
             <div class="list-group-item">
                 @Item(item)
@@ -21,7 +21,7 @@ else
 }
 
 @code {
-    IEnumerable<TItem> items;
+    IEnumerable<TItem> _items = null!;
 
     [Parameter] public Func<Task<IEnumerable<TItem>>> Loader { get; set; }
     [Parameter] public RenderFragment Loading { get; set; }
@@ -31,6 +31,9 @@ else
 
     protected override async Task OnParametersSetAsync()
     {
-        items = await Loader();
+        if (Loader is not null)
+        {
+            _items = await Loader();
+        }
     }
 }
