@@ -1,13 +1,8 @@
-﻿namespace BlazingPizza.ComponentsLibrary.Map
+﻿namespace BlazingPizza.ComponentsLibrary.Map;
+
+public class Marker : Point
 {
-    public class Marker
-    {
-        public string Description { get; set; }
+    public string Description { get; set; }
 
-        public double X { get; set; }
-
-        public double Y { get; set; }
-
-        public bool ShowPopup { get; set; }
-    }
+    public bool ShowPopup { get; set; }
 }
