@@ -1,133 +1,133 @@
-﻿namespace BlazingPizza.Server
+﻿namespace BlazingPizza.Server;
+
+public static class SeedData
 {
-    public static class SeedData
+    public static Task InitializeAsync(PizzaStoreContext db)
     {
-        public static void Initialize(PizzaStoreContext db)
+        var toppings = new Topping[]
         {
-            var toppings = new Topping[]
-            {
-                new Topping()
+                new Topping
                 {
                     Name = "Extra cheese",
                     Price = 2.50m,
                 },
-                new Topping()
+                new Topping
                 {
                     Name = "American bacon",
                     Price = 2.99m,
                 },
-                new Topping()
+                new Topping
                 {
                     Name = "British bacon",
                     Price = 2.99m,
                 },
-                new Topping()
+                new Topping
                 {
                     Name = "Canadian bacon",
                     Price = 2.99m,
                 },
-                new Topping()
+                new Topping
                 {
                     Name = "Tea and crumpets",
                     Price = 5.00m
                 },
-                new Topping()
+                new Topping
                 {
                     Name = "Fresh-baked scones",
                     Price = 4.50m,
                 },
-                new Topping()
+                new Topping
                 {
                     Name = "Bell peppers",
                     Price = 1.00m,
                 },
-                new Topping()
+                new Topping
                 {
                     Name = "Onions",
                     Price = 1.00m,
                 },
-                new Topping()
+                new Topping
                 {
                     Name = "Mushrooms",
                     Price = 1.00m,
                 },
-                new Topping()
+                new Topping
                 {
                     Name = "Pepperoni",
                     Price = 1.00m,
                 },
-                new Topping()
+                new Topping
                 {
                     Name = "Duck sausage",
                     Price = 3.20m,
                 },
-                new Topping()
+                new Topping
                 {
                     Name = "Venison meatballs",
                     Price = 2.50m,
                 },
-                new Topping()
+                new Topping
                 {
                     Name = "Served on a silver platter",
                     Price = 250.99m,
                 },
-                new Topping()
+                new Topping
                 {
                     Name = "Lobster on top",
                     Price = 64.50m,
                 },
-                new Topping()
+                new Topping
                 {
                     Name = "Sturgeon caviar",
                     Price = 101.75m,
                 },
-                new Topping()
+                new Topping
                 {
                     Name = "Artichoke hearts",
                     Price = 3.40m,
                 },
-                new Topping()
+                new Topping
                 {
                     Name = "Fresh tomatoes",
                     Price = 1.50m,
                 },
-                new Topping()
+                new Topping
                 {
                     Name = "Basil",
                     Price = 1.50m,
                 },
-                new Topping()
+                new Topping
                 {
                     Name = "Steak (medium-rare)",
                     Price = 8.50m,
                 },
-                new Topping()
+                new Topping
                 {
                     Name = "Blazing hot peppers",
                     Price = 4.20m,
                 },
-                new Topping()
+                new Topping
                 {
                     Name = "Buffalo chicken",
                     Price = 5.00m,
                 },
-                new Topping()
+                new Topping
                 {
                     Name = "Blue cheese",
                     Price = 2.50m,
                 },
-            };
+        };
 
-            var specials = new PizzaSpecial[]
-            {
-                new PizzaSpecial()
+        var specials = new PizzaSpecial[]
+        {
+                new PizzaSpecial
                 {
                     Name = "Basic Cheese Pizza",
                     Description = "It's cheesy and delicious. Why wouldn't you want one?",
                     BasePrice = 9.99m,
                     ImageUrl = "img/pizzas/cheese.jpg",
                 },
-                new PizzaSpecial()
+                new PizzaSpecial
                 {
                     Id = 2,
                     Name = "The Baconatorizor",
@@ -135,7 +135,7 @@
                     BasePrice = 11.99m,
                     ImageUrl = "img/pizzas/bacon.jpg",
                 },
-                new PizzaSpecial()
+                new PizzaSpecial
                 {
                     Id = 3,
                     Name = "Classic pepperoni",
@@ -143,7 +143,7 @@
                     BasePrice = 10.50m,
                     ImageUrl = "img/pizzas/pepperoni.jpg",
                 },
-                new PizzaSpecial()
+                new PizzaSpecial
                 {
                     Id = 4,
                     Name = "Buffalo chicken",
@@ -151,7 +151,7 @@
                     BasePrice = 12.75m,
                     ImageUrl = "img/pizzas/meaty.jpg",
                 },
-                new PizzaSpecial()
+                new PizzaSpecial
                 {
                     Id = 5,
                     Name = "Mushroom Lovers",
@@ -159,7 +159,7 @@
                     BasePrice = 11.00m,
                     ImageUrl = "img/pizzas/mushroom.jpg",
                 },
-                new PizzaSpecial()
+                new PizzaSpecial
                 {
                     Id = 6,
                     Name = "The Brit",
@@ -167,7 +167,7 @@
                     BasePrice = 10.25m,
                     ImageUrl = "img/pizzas/brit.jpg",
                 },
-                new PizzaSpecial()
+                new PizzaSpecial
                 {
                     Id = 7,
                     Name = "Veggie Delight",
@@ -175,7 +175,7 @@
                     BasePrice = 11.50m,
                     ImageUrl = "img/pizzas/salad.jpg",
                 },
-                new PizzaSpecial()
+                new PizzaSpecial
                 {
                     Id = 8,
                     Name = "Margherita",
@@ -183,11 +183,11 @@
                     BasePrice = 9.99m,
                     ImageUrl = "img/pizzas/margherita.jpg",
                 },
-            };
+        };
 
-            db.Toppings.AddRange(toppings);
-            db.Specials.AddRange(specials);
-            db.SaveChanges();
-        }
+        db.Toppings.AddRange(toppings);
+        db.Specials.AddRange(specials);
+
+        return db.SaveChangesAsync();
     }
 }
