@@ -1,25 +1,20 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿namespace BlazingPizza;
 
-namespace BlazingPizza
+public class Order
 {
-    public class Order
-    {
-        public int OrderId { get; set; }
+    public int OrderId { get; set; }
 
-        public string UserId { get; set; }
+    public string UserId { get; set; }
 
-        public DateTime CreatedTime { get; set; }
+    public DateTime CreatedTime { get; set; }
 
-        public Address DeliveryAddress { get; set; } = new Address();
+    public Address DeliveryAddress { get; set; } = new();
 
-        public LatLong DeliveryLocation { get; set; }
+    public LatLong DeliveryLocation { get; set; }
 
-        public List<Pizza> Pizzas { get; set; } = new List<Pizza>();
+    public List<Pizza> Pizzas { get; set; } = new();
 
-        public decimal GetTotalPrice() => Pizzas.Sum(p => p.GetTotalPrice());
+    public decimal GetTotalPrice() => Pizzas.Sum(p => p.GetTotalPrice());
 
-        public string GetFormattedTotalPrice() => GetTotalPrice().ToString("0.00");
-    }
+    public string GetFormattedTotalPrice() => GetTotalPrice().ToString("0.00");
 }
