@@ -1,79 +1,85 @@
 ﻿using BlazingPizza.ComponentsLibrary.Map;
-using System;
-using System.Collections.Generic;
 
-namespace BlazingPizza
+namespace BlazingPizza;
+
+public class OrderWithStatus
 {
-    public class OrderWithStatus
+    public readonly static TimeSpan PreparationDuration = TimeSpan.FromSeconds(10);
+    public readonly static TimeSpan DeliveryDuration = TimeSpan.FromMinutes(1); // Unrealistic, but more interesting to watch
+
+    public Order Order { get; set; }
+
+    public string StatusText { get; set; }
+
+    public bool IsDelivered => StatusText == "Delivered";
+
+    public List<Marker> MapMarkers { get; set; }
+
+    public static OrderWithStatus FromOrder(Order order)
     {
-        public readonly static TimeSpan PreparationDuration = TimeSpan.FromSeconds(10);
-        public readonly static TimeSpan DeliveryDuration = TimeSpan.FromMinutes(1); // Unrealistic, but more interesting to watch
+        // To simulate a real backend process, we fake status updates based on the amount
+        // of time since the order was placed
+        string statusText;
+        List<Marker> mapMarkers;
+        var dispatchTime = order.CreatedTime.Add(PreparationDuration);
 
-        public Order Order { get; set; }
-
-        public string StatusText { get; set; }
-
-        public bool IsDelivered => StatusText == "Delivered";
-
-        public List<Marker> MapMarkers { get; set; }
-
-        public static OrderWithStatus FromOrder(Order order)
+        if (DateTime.Now < dispatchTime)
         {
-            // To simulate a real backend process, we fake status updates based on the amount
-            // of time since the order was placed
-            string statusText;
-            List<Marker> mapMarkers;
-            var dispatchTime = order.CreatedTime.Add(PreparationDuration);
-
-            if (DateTime.Now < dispatchTime)
-            {
-                statusText = "Preparing";
-                mapMarkers = new List<Marker>
+            statusText = "Preparing";
+            mapMarkers = new List<Marker>
                 {
                     ToMapMarker("You", order.DeliveryLocation, showPopup: true)
                 };
-            }
-            else if (DateTime.Now < dispatchTime + DeliveryDuration)
-            {
-                statusText = "Out for delivery";
+        }
+        else if (DateTime.Now < dispatchTime + DeliveryDuration)
+        {
+            statusText = "Out for delivery";
 
-                var startPosition = ComputeStartPosition(order);
-                var proportionOfDeliveryCompleted = Math.Min(1, (DateTime.Now - dispatchTime).TotalMilliseconds / DeliveryDuration.TotalMilliseconds);
-                var driverPosition = LatLong.Interpolate(startPosition, order.DeliveryLocation, proportionOfDeliveryCompleted);
-                mapMarkers = new List<Marker>
+            var startPosition = ComputeStartPosition(order);
+            var proportionOfDeliveryCompleted = Math.Min(1, (DateTime.Now - dispatchTime).TotalMilliseconds / DeliveryDuration.TotalMilliseconds);
+            var driverPosition = LatLong.Interpolate(startPosition, order.DeliveryLocation, proportionOfDeliveryCompleted);
+            mapMarkers = new List<Marker>
                 {
                     ToMapMarker("You", order.DeliveryLocation),
                     ToMapMarker("Driver", driverPosition, showPopup: true),
                 };
-            }
-            else
-            {
-                statusText = "Delivered";
-                mapMarkers = new List<Marker>
+        }
+        else
+        {
+            statusText = "Delivered";
+            mapMarkers = new List<Marker>
                 {
                     ToMapMarker("Delivery location", order.DeliveryLocation, showPopup: true),
                 };
-            }
-
-            return new OrderWithStatus
-            {
-                Order = order,
-                StatusText = statusText,
-                MapMarkers = mapMarkers,
-            };
         }
 
-        private static LatLong ComputeStartPosition(Order order)
+        return new OrderWithStatus
         {
-            // Random but deterministic based on order ID
-            var rng = new Random(order.OrderId);
-            var distance = 0.01 + rng.NextDouble() * 0.02;
-            var angle = rng.NextDouble() * Math.PI * 2;
-            var offset = (distance * Math.Cos(angle), distance * Math.Sin(angle));
-            return new LatLong(order.DeliveryLocation.Latitude + offset.Item1, order.DeliveryLocation.Longitude + offset.Item2);
-        }
-
-        static Marker ToMapMarker(string description, LatLong coords, bool showPopup = false)
-            => new Marker { Description = description, X = coords.Longitude, Y = coords.Latitude, ShowPopup = showPopup };
+            Order = order,
+            StatusText = statusText,
+            MapMarkers = mapMarkers,
+        };
     }
+
+    private static LatLong ComputeStartPosition(Order order)
+    {
+        // Random but deterministic based on order ID
+        var rng = new Random(order.OrderId);
+        var distance = 0.01 + rng.NextDouble() * 0.02;
+        var angle = rng.NextDouble() * Math.PI * 2;
+        var offset = (distance * Math.Cos(angle), distance * Math.Sin(angle));
+        
+        return new LatLong(
+            order.DeliveryLocation?.Latitude ?? 0 + offset.Item1,
+            order.DeliveryLocation?.Longitude ?? 0 + offset.Item2);
+    }
+
+    static Marker ToMapMarker(string description, LatLong coords, bool showPopup = false) =>
+        new()
+        {
+            Description = description,
+            X = coords?.Longitude ?? 0,
+            Y = coords?.Latitude ?? 0,
+            ShowPopup = showPopup
+        };
 }
