@@ -1,20 +1,19 @@
-﻿namespace BlazingPizza
+﻿namespace BlazingPizza;
+
+/// <summary>
+/// Represents a pre-configured template for a pizza a user can order
+/// </summary>
+public class PizzaSpecial
 {
-    /// <summary>
-    /// Represents a pre-configured template for a pizza a user can order
-    /// </summary>
-    public class PizzaSpecial
-    {
-        public int Id { get; set; }
+    public int Id { get; set; }
 
-        public string Name { get; set; }
+    public string Name { get; set; }
 
-        public decimal BasePrice { get; set; }
+    public decimal BasePrice { get; set; }
 
-        public string Description { get; set; }
+    public string Description { get; set; }
 
-        public string ImageUrl { get; set; }
+    public string ImageUrl { get; set; }
 
-        public string GetFormattedBasePrice() => BasePrice.ToString("0.00");
-    }
+    public string GetFormattedBasePrice() => BasePrice.ToString("0.00");
 }
