@@ -1,13 +1,9 @@
-﻿using Microsoft.JSInterop;
-using System.Threading.Tasks;
+﻿namespace BlazingPizza.Client;
 
-namespace BlazingPizza.Client
+public static class JSRuntimeExtensions
 {
-    public static class JSRuntimeExtensions
+    public static ValueTask<bool> Confirm(this IJSRuntime jsRuntime, string message)
     {
-        public static ValueTask<bool> Confirm(this IJSRuntime jsRuntime, string message)
-        {
-            return jsRuntime.InvokeAsync<bool>("confirm", message);
-        }
+        return jsRuntime.InvokeAsync<bool>("confirm", message);
     }
 }
