@@ -13,7 +13,7 @@
             <div class="col">
                 <h5>@item.Order.CreatedTime.ToLongDateString()</h5>
                 Items:
-                <strong>@item.Order.Pizzas.Count()</strong>;
+                <strong>@item.Order.Pizzas.Count</strong>;
                 Total price:
                 <strong>£@item.Order.GetFormattedTotalPrice()</strong>
             </div>
