@@ -8,7 +8,7 @@
         }
     </ul>
     <div class="item-price">
-        @Pizza.GetFormattedTotalPrice()
+        @Pizza?.GetFormattedTotalPrice()
     </div>
 </div>
 
