@@ -5,9 +5,9 @@
 
 @code {
     string elementId = $"map-{Guid.NewGuid().ToString("D")}";
-    
+
     [Parameter] public double Zoom { get; set; }
-    [Parameter] public List<Marker> Markers { get; set; }
+    [Parameter] public List<Marker> Markers { get; set; } = default!;
 
     protected async override Task OnAfterRenderAsync(bool firstRender)
     {
