@@ -1,26 +1,16 @@
-﻿using Microsoft.AspNetCore.ApiAuthorization.IdentityServer;
-using Microsoft.AspNetCore.Mvc;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿namespace BlazingPizza.Server;
 
-namespace BlazingPizza.Server
+public class OidcConfigurationController : Controller
 {
-    public class OidcConfigurationController : Controller
+    public OidcConfigurationController(IClientRequestParametersProvider clientRequestParametersProvider) =>
+        ClientRequestParametersProvider = clientRequestParametersProvider;
+
+    public IClientRequestParametersProvider ClientRequestParametersProvider { get; }
+
+    [HttpGet("_configuration/{clientId}")]
+    public IActionResult GetClientRequestParameters([FromRoute] string clientId)
     {
-        public OidcConfigurationController(IClientRequestParametersProvider clientRequestParametersProvider)
-        {
-            ClientRequestParametersProvider = clientRequestParametersProvider;
-        }
-
-        public IClientRequestParametersProvider ClientRequestParametersProvider { get; }
-
-        [HttpGet("_configuration/{clientId}")]
-        public IActionResult GetClientRequestParameters([FromRoute]string clientId)
-        {
-            var parameters = ClientRequestParametersProvider.GetClientParameters(HttpContext, clientId);
-            return Ok(parameters);
-        }
+        var parameters = ClientRequestParametersProvider.GetClientParameters(HttpContext, clientId);
+        return Ok(parameters);
     }
 }
