@@ -1,16 +1,10 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿namespace BlazingPizza.Server;
 
-namespace BlazingPizza.Server
+[Route("pizzas")]
+[ApiController]
+public class PizzasController : Controller
 {
-    [Route("pizzas")]
-    [ApiController]
-    public class PizzasController : Controller
-    {
-        private readonly PizzaStoreContext _db;
+    private readonly PizzaStoreContext _db;
 
-        public PizzasController(PizzaStoreContext db)
-        {
-            _db = db;
-        }
-    }
+    public PizzasController(PizzaStoreContext db) => _db = db;
 }
