@@ -1,26 +1,19 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
+﻿namespace BlazingPizza.Server;
 
-namespace BlazingPizza.Server
+[Route("specials")]
+[ApiController]
+public class SpecialsController : Controller
 {
-    [Route("specials")]
-    [ApiController]
-    public class SpecialsController : Controller
+    private readonly PizzaStoreContext _db;
+
+    public SpecialsController(PizzaStoreContext db)
     {
-        private readonly PizzaStoreContext _db;
+        _db = db;
+    }
 
-        public SpecialsController(PizzaStoreContext db)
-        {
-            _db = db;
-        }
-
-        [HttpGet]
-        public async Task<ActionResult<List<PizzaSpecial>>> GetSpecials()
-        {
-            return (await _db.Specials.ToListAsync()).OrderByDescending(s => s.BasePrice).ToList();
-        }
+    [HttpGet]
+    public async Task<ActionResult<List<PizzaSpecial>>> GetSpecials()
+    {
+        return (await _db.Specials.ToListAsync()).OrderByDescending(s => s.BasePrice).ToList();
     }
 }
