@@ -1,10 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
-
-namespace BlazingPizza.Server
+﻿namespace BlazingPizza.Server
 {
     [Route("toppings")]
     [ApiController]
